Use explicit args for shfmt
shfmt provides different behavior locally vs in the project and it took
me a few minutes to get the write explicit command for formatting.

This marks the shfmt as the exact args needed so it overwrites any other
system behaviors during development and testing and the ci system will
be self explanatory for the full args needed.